src/requirements.txt
# Основные зависимости здесь
fastapi==0.115.11
uvicorn==0.34.0
alembic==1.14.1
sqlalchemy==2.0.38
asyncpg==0.30.0
pydantic-settings==2.7.0
fastapi-users[sqlalchemy]==14.0.1
loguru==0.7.3
click==8.1.8
pydantic-extra-types[phone-numbers]==2.10.2
phonenumbers==8.13.55
httpx>=0.26.0,<0.29.0
python-dotenv==1.0.1
python-telegram-bot==20.8
redis==5.2.1
sqladmin==0.20.1
pandas==2.2.3
